fix: validate transaction money values as decimals

+++ backend/app/schemas/transaction.py
@@ -1,4 +1,5 @@
 from datetime import datetime
+from decimal import Decimal
 from typing import Literal
 
 from pydantic import BaseModel, Field
@@ -6,7 +7,7 @@ from pydantic import BaseModel, Field
 
 class TransactionCreate(BaseModel):
     title: str = Field(min_length=1, max_length=100)
-    amount: float = Field(gt=0)
+    amount: Decimal = Field(gt=0, max_digits=14, decimal_places=2)
     type: Literal["income", "expense"]
     category: str = Field(min_length=1, max_length=50)
 
@@ -14,7 +15,7 @@ class TransactionCreate(BaseModel):
 class TransactionResponse(BaseModel):
     id: int
     title: str
-    amount: float
+    amount: Decimal
     type: str
     category: str
     created_at: datetime
